python_engine/requirements.txt
# BioMapper Python Engine Dependencies
# Core AI/ML Libraries
transformers>=4.30.0
torch>=2.0.0
tensorflow>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scikit-bio>=0.5.8
accelerate>=0.20.0
datasets>=2.12.0
sentencepiece>=0.1.99
safetensors>=0.3.0
einops>=0.6.0
captum>=0.6.0
joblib>=1.3.0

# Bioinformatics Libraries
biopython>=1.81
dendropy>=4.5.2
ete3>=3.1.2

# Quantum Computing (Optional - for quantum benchmarks)
qiskit>=0.43.0
qiskit_algorithms>=0.2.0
qiskit-optimization>=0.5.0
qiskit-ibm-runtime>=0.11.0
qiskit_aer>=0.12.0

# Web Framework & API
flask>=2.3.0
flask-cors>=4.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=11.0.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Interface (Optional)
streamlit>=1.24.0

# HTTP Requests
requests>=2.31.0

# Note: sqlite3 is built into Python 3.x
# Note: ClustalW and MAFFT are external tools, install separately if needed